Documentation + Fixes

+++ nutrifit-backend/controller/favoriteController.js
@@ -28,10 +28,10 @@ export async function favoriteRecipe(req, res) {
         // Add to favorites
         await connection.promise().query(
             `INSERT INTO NutriFit.user_favorites 
-                (favorite_id, user_id, recipe_id)
-            VALUES (?, ?, ?)`,
+                (user_id, recipe_id)
+            VALUES 
+                (?, ?)`,
             [
-                uuidv4(),
                 userId,
                 recipeId
             ]
@@ -65,7 +65,7 @@ export async function getUserFavorites(req, res) {
             return res.status(404).json({ body: 'User has no favorites' });
         }
 
-        return res.status(200).json({ favorites: favorites[0] })
+        return res.status(200).json({ favorites: favorites })
     } catch (err) {
         console.log(err);
         return res.status(500).json({ error: 'Something unexpected happened' });
@@ -75,9 +75,21 @@ export async function getUserFavorites(req, res) {
 export async function removeFavorite(req, res) {
     try {
         const connection = await mysql.createConnection(connectionString);
+
+        const decodedToken = await authentication(req);
+
+        const userId = decodedToken.sub;
+        const userVerification = await connection.promise().query(
+            'SELECT * FROM NutriFit.user WHERE user_id = ?',
+            [userId]
+        );
+        if (!userVerification[0] || userVerification[0].length === 0) {
+            return res.status(400).json({ error: 'User not found' });
+        }
+
         await connection.promise().query(
             'DELETE FROM NutriFit.user_favorites WHERE user_id = ? AND recipe_id = ?',
-            [req.body.user_id, req.body.recipe_id]
+            [userId, req.body.recipe_id]
         );
 
         return res.status(200).json({ message: 'Recipe removed from favorites successfully' });
